Properly sync the contents tree when using remote control.
Although measures already were in place to prevent a race condition of kinds (in HelpPage::acceptNavigationRequest() and HelpViewer::source()), testing has shown that these measures are not sufficient. This is because in timings that I did, the order is the following:

At t0: RemoteControl::handleSetSourceCommand() causes a call to HelpViewer::setSource() which does QWebEngineView::load(). WebEngine goes off to do its thing.
At t0+163ms: RemoteControl::handleSyncContentsCommand() asks the main window to sync the contents tree view.
At t0+163ms (immediately after): HelpPage::source() is called in which m_loadingUrl was evaluated, which was expected to contain the url set at t0, but is not set until 120ms later:
At t0+284ms: WebEngine signals loadStarted(), HelpPage::acceptNavigationRequest() gets called, in which the URL was stored in m_loadingUrl to counter the race condition.

This is now countered by storing m_loadingUrl in HelpViewer::setSource(). This way, WebEngine can take as longs as it wants to fire up and render all images etc, and the contents tree view will have synced already.

Obviously, without this change, the tree view will always sync with the previously set source, not the current one (with the above timings at least).

With this, the corresponding code in HelpPage::acceptNavigationRequest() has become unnecessary, which may make overriding it obsolete I don't know, I haven't looked at the base implementation.